% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBathyTransect_function.R
\name{getBathyTransect}
\alias{getBathyTransect}
\title{getBathyTransect}
\usage{
getBathyTransect(bathymetry = NULL, startFence = NULL, endFence = NULL)
}
\arguments{
\item{bathymetry}{An object of class Bathymetry as defiend by the marmap package. Defaults to NULL. If not specified, a bathymetry object will be called form transect parameters}

\item{startFence}{An vector containing the longitude and latitude points denoting the begining of the fence.}

\item{startFence}{An vector containing the longitude and latitude points denoting the end of the fence.}
}
\description{
This function returns a transect from a bathymetry object. if a bathymetry object hasn't been specified, it first querys the NOAA database to get one. If bathymetry data is provided but no start or end to the transect has been specified, an interactive map allows users to select transect start and end points
}
\examples{
getBathyTransect(startFence = c(-157.68330, 21.28333), endFence = c(-157.53330, 21.41667))
}
\keyword{and}
\keyword{bathymetry}
\keyword{get}
\keyword{transect}
