% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotonuseToLeg_function.R
\name{hypotonuseToLeg}
\alias{hypotonuseToLeg}
\title{hypotonuseToLeg}
\usage{
hypotonuseToLeg(x)
}
\arguments{
\item{x}{The length of a 45-45-90 triangle's leg}
}
\description{
This function returns calculate the length of one leg of a 45° 45° 90° triangle from its hypotenuse
}
\examples{
hypotonuseToLeg()
}
\keyword{hypotonuseToLeg}
