% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSensorFenceR_function.R
\name{runSensorFenceR}
\alias{runSensorFenceR}
\title{runSensorFenceR}
\usage{
runSensorFenceR(bathymetry = NULL, fenceStart = NULL, fenceEnd = NULL,
  upperDepthLimit = NULL, lowerDepthLimit = NULL, heightOfSensor = NULL,
  minimumDetectionThreshold = NULL, halfOfminimumDetectionThreshold = NULL,
  numberOfSensorsToPlace = NULL, minimumFenceHeight = 150,
  swimSpeedOfStudySpecies = 0.5, plot_results = TRUE)
}
\arguments{
\item{bathymetry}{Inherited from the parent function}

\item{fenceStart}{Inherited from the parent function}

\item{fenceEnd}{Inherited from the parent function}

\item{upperDepthLimit}{The shallowest depth at which to place a sensor. Note, depth is a negative number.}

\item{lowerDepthLimit}{The deepest depth at which to place a sensor. Note, depth is a negative number.}

\item{heightOfSensor}{The distance off the bottom (in meters) that the sensor is positioned}

\item{minimumDetectionThreshold}{The radius (in meters) from the sensor where 1/2 of the minimum acceptable detection range is achieved. This is generally determined through range testing.}

\item{halfOfminimumDetectionThreshold}{The radius (in meters) from the sensor where 1/2 of the minimum acceptable detection range is achieved. For example, if the minimum allowable detection percentage is 50\%, this would be the distance from the sensor at which the detection threshold is 25\%. This is typically determined from range testing.}

\item{numberOfSensorsToPlace}{The number of sensors available from which to construct the fence line (integer)}

\item{minimumFenceHeight}{The height (m) off the seafloor that a transmisson must be detectable. Leave NULL if not specified.}

\item{swimSpeedOfStudySpecies}{# The speed at which the transmitter may cross the sensor fence (m/s)}
}
\description{
This Function Places Static Sensors Along a Linear Transect
}
\examples{
runSensorFenceR(numberOfSensorsToPlace = 10, fenceStart = c(-157.68330, 21.28333), fenceEnd = c(-157.53330, 21.41667), upperDepthLimit = -100, lowerDepthLimit = -500, minimumDetectionThreshold = 500, halfOfminimumDetectionThreshold = 650, minimumFenceHeight = 100, swimSpeedOfStudySpecies = .5)
}
\keyword{Place}
\keyword{a}
\keyword{along}
\keyword{linear}
\keyword{sensors}
\keyword{static}
\keyword{transect}
